Feat/ssao (#11)
* wip : ssao looks bad

* feqt: sseo now looks fine, still need to do blur

* feat : added blur
todo clean and merge and do something to turn off / on passes

* added label + enabling passes on rendergraph
todo : should be resetting inputs

* added res manager for the rdg

* cleaned a bit more, preparing for rebase

* fix issues

+++ PyriteCore/src/world/camera.cpp
@@ -157,7 +157,7 @@ bool Frustum::isOnOrForwardPlan(const Plane &plane, const AABB &boundingBox)
   return -r <= plane.signedDistanceTo(center);
 }
 
-void FreecamController::processUserInputs(float deltaTime)
+void FreecamController::processUserInputs(double deltaTime)
 {
   if (!m_camera) return;
